A separate path created in this branch for a new user to be added
This allows a non-gen user create an account and login. The process
takes all User fields in the login.jsp View in a New User adjacent form
to the original Login form, then passes it to a CreateLogin Controller,
which, in turn creates an instance of a CreateUser Service, which calls
the UserArrayList Singleton Object's array of Users, and adds the new
User, before returning a boolean true once created. The CreateLogin
Controller should then redirect to the Login Screen, and populate the
Login form's username field.

Presently this last feay=ture doesn't work. Will fix this, then look at:
-JS client-validation;
-Autogen password and sending to each new user created, and adding code
so they can change this password once logged in;
-Build JUnit Tests;
-Test Scenario walkthrough for holes in logic;
ALSO:
-CSS styling, and BootStrap UI integration;