Modified Input fields email and password of 'Create New User Form' on login.jsp View.
The respective fields were modified to allow for some Client-side
validation. The 'email' input type was set so that it does the HTML5
email validation automatically. I then and included a HTML5 Regex for
the Password, so that it only accepts ones of 6 characters or greater,
and with both upper and lowercase and digits, displaying a custom
warning otherwise.